+++ lib/Dialect/AIE/IR/AIETargetModel.cpp
@@ -64,33 +64,32 @@ std::optional<TileID> AIE1TargetModel::getMemSouth(TileID src) const {
   return ret;
 }
 
-bool AIE1TargetModel::isMemWest(uint32_t srcCol, uint32_t srcRow,
-                                uint32_t dstCol, uint32_t dstRow) const {
+bool AIE1TargetModel::isMemWest(int srcCol, int srcRow, int dstCol,
+                                int dstRow) const {
   bool IsEvenRow = ((srcRow % 2) == 0);
   return (IsEvenRow && isInternal(srcCol, srcRow, dstCol, dstRow)) ||
          (!IsEvenRow && isWest(srcCol, srcRow, dstCol, dstRow));
 }
 
-bool AIE1TargetModel::isMemEast(uint32_t srcCol, uint32_t srcRow,
-                                uint32_t dstCol, uint32_t dstRow) const {
+bool AIE1TargetModel::isMemEast(int srcCol, int srcRow, int dstCol,
+                                int dstRow) const {
   bool IsEvenRow = ((srcRow % 2) == 0);
   return (!IsEvenRow && isInternal(srcCol, srcRow, dstCol, dstRow)) ||
          (IsEvenRow && isEast(srcCol, srcRow, dstCol, dstRow));
 }
 
-bool AIE1TargetModel::isMemNorth(uint32_t srcCol, uint32_t srcRow,
-                                 uint32_t dstCol, uint32_t dstRow) const {
+bool AIE1TargetModel::isMemNorth(int srcCol, int srcRow, int dstCol,
+                                 int dstRow) const {
   return isNorth(srcCol, srcRow, dstCol, dstRow);
 }
 
-bool AIE1TargetModel::isMemSouth(uint32_t srcCol, uint32_t srcRow,
-                                 uint32_t dstCol, uint32_t dstRow) const {
+bool AIE1TargetModel::isMemSouth(int srcCol, int srcRow, int dstCol,
+                                 int dstRow) const {
   return isSouth(srcCol, srcRow, dstCol, dstRow);
 }
 
-bool AIE1TargetModel::isLegalMemAffinity(uint32_t coreCol, uint32_t coreRow,
-                                         uint32_t memCol,
-                                         uint32_t memRow) const {
+bool AIE1TargetModel::isLegalMemAffinity(int coreCol, int coreRow, int memCol,
+                                         int memRow) const {
   bool IsEvenRow = ((coreRow % 2) == 0);
 
   bool IsMemWest = (isWest(coreCol, coreRow, memCol, memRow) && !IsEvenRow) ||
@@ -105,7 +104,7 @@ bool AIE1TargetModel::isLegalMemAffinity(uint32_t coreCol, uint32_t coreRow,
   return IsMemSouth || IsMemNorth || IsMemWest || IsMemEast;
 }
 uint32_t
-AIE1TargetModel::getNumDestSwitchboxConnections(uint32_t col, uint32_t row,
+AIE1TargetModel::getNumDestSwitchboxConnections(int col, int row,
                                                 WireBundle bundle) const {
   if (isShimNOCTile(col, row) || isShimPLTile(col, row))
     switch (bundle) {
@@ -158,7 +157,7 @@ AIE1TargetModel::getNumDestSwitchboxConnections(uint32_t col, uint32_t row,
     }
 }
 uint32_t
-AIE1TargetModel::getNumSourceSwitchboxConnections(uint32_t col, uint32_t row,
+AIE1TargetModel::getNumSourceSwitchboxConnections(int col, int row,
                                                   WireBundle bundle) const {
   if (isShimNOCTile(col, row) || isShimPLTile(col, row))
     switch (bundle) {
@@ -215,7 +214,7 @@ AIE1TargetModel::getNumSourceSwitchboxConnections(uint32_t col, uint32_t row,
     }
 }
 uint32_t
-AIE1TargetModel::getNumDestShimMuxConnections(uint32_t col, uint32_t row,
+AIE1TargetModel::getNumDestShimMuxConnections(int col, int row,
                                               WireBundle bundle) const {
   if (isShimNOCorPLTile(col, row))
     switch (bundle) {
@@ -234,7 +233,7 @@ AIE1TargetModel::getNumDestShimMuxConnections(uint32_t col, uint32_t row,
     return 0;
 }
 uint32_t
-AIE1TargetModel::getNumSourceShimMuxConnections(uint32_t col, uint32_t row,
+AIE1TargetModel::getNumSourceShimMuxConnections(int col, int row,
                                                 WireBundle bundle) const {
   if (isShimNOCTile(col, row))
     switch (bundle) {
@@ -254,9 +253,9 @@ AIE1TargetModel::getNumSourceShimMuxConnections(uint32_t col, uint32_t row,
 }
 
 bool AIE1TargetModel::isLegalMemtileConnection(WireBundle srcBundle,
-                                               uint32_t srcChan,
+                                               int srcChan,
                                                WireBundle dstBundle,
-                                               uint32_t dstChan) const {
+                                               int dstChan) const {
   return false;
 }
 
@@ -296,29 +295,28 @@ std::optional<TileID> AIE2TargetModel::getMemSouth(TileID src) const {
   return ret;
 }
 
-bool AIE2TargetModel::isMemWest(uint32_t srcCol, uint32_t srcRow,
-                                uint32_t dstCol, uint32_t dstRow) const {
+bool AIE2TargetModel::isMemWest(int srcCol, int srcRow, int dstCol,
+                                int dstRow) const {
   return isWest(srcCol, srcRow, dstCol, dstRow);
 }
 
-bool AIE2TargetModel::isMemEast(uint32_t srcCol, uint32_t srcRow,
-                                uint32_t dstCol, uint32_t dstRow) const {
+bool AIE2TargetModel::isMemEast(int srcCol, int srcRow, int dstCol,
+                                int dstRow) const {
   return isInternal(srcCol, srcRow, dstCol, dstRow);
 }
 
-bool AIE2TargetModel::isMemNorth(uint32_t srcCol, uint32_t srcRow,
-                                 uint32_t dstCol, uint32_t dstRow) const {
+bool AIE2TargetModel::isMemNorth(int srcCol, int srcRow, int dstCol,
+                                 int dstRow) const {
   return isNorth(srcCol, srcRow, dstCol, dstRow);
 }
 
-bool AIE2TargetModel::isMemSouth(uint32_t srcCol, uint32_t srcRow,
-                                 uint32_t dstCol, uint32_t dstRow) const {
+bool AIE2TargetModel::isMemSouth(int srcCol, int srcRow, int dstCol,
+                                 int dstRow) const {
   return isSouth(srcCol, srcRow, dstCol, dstRow);
 }
 
-bool AIE2TargetModel::isLegalMemAffinity(uint32_t coreCol, uint32_t coreRow,
-                                         uint32_t memCol,
-                                         uint32_t memRow) const {
+bool AIE2TargetModel::isLegalMemAffinity(int coreCol, int coreRow, int memCol,
+                                         int memRow) const {
 
   bool IsMemWest = isMemWest(coreCol, coreRow, memCol, memRow);
   bool IsMemEast = isMemEast(coreCol, coreRow, memCol, memRow);
@@ -334,7 +332,7 @@ bool AIE2TargetModel::isLegalMemAffinity(uint32_t coreCol, uint32_t coreRow,
            IsMemWest || IsMemEast;
 }
 uint32_t
-AIE2TargetModel::getNumDestSwitchboxConnections(uint32_t col, uint32_t row,
+AIE2TargetModel::getNumDestSwitchboxConnections(int col, int row,
                                                 WireBundle bundle) const {
   if (isMemTile(col, row))
     switch (bundle) {
@@ -398,7 +396,7 @@ AIE2TargetModel::getNumDestSwitchboxConnections(uint32_t col, uint32_t row,
     }
 }
 uint32_t
-AIE2TargetModel::getNumSourceSwitchboxConnections(uint32_t col, uint32_t row,
+AIE2TargetModel::getNumSourceSwitchboxConnections(int col, int row,
                                                   WireBundle bundle) const {
   if (isMemTile(col, row))
     switch (bundle) {
@@ -469,7 +467,7 @@ AIE2TargetModel::getNumSourceSwitchboxConnections(uint32_t col, uint32_t row,
     }
 }
 uint32_t
-AIE2TargetModel::getNumDestShimMuxConnections(uint32_t col, uint32_t row,
+AIE2TargetModel::getNumDestShimMuxConnections(int col, int row,
                                               WireBundle bundle) const {
   if (isShimNOCorPLTile(col, row))
     switch (bundle) {
@@ -488,7 +486,7 @@ AIE2TargetModel::getNumDestShimMuxConnections(uint32_t col, uint32_t row,
     return 0;
 }
 uint32_t
-AIE2TargetModel::getNumSourceShimMuxConnections(uint32_t col, uint32_t row,
+AIE2TargetModel::getNumSourceShimMuxConnections(int col, int row,
                                                 WireBundle bundle) const {
   if (isShimNOCTile(col, row))
     switch (bundle) {
@@ -508,9 +506,9 @@ AIE2TargetModel::getNumSourceShimMuxConnections(uint32_t col, uint32_t row,
 }
 
 bool AIE2TargetModel::isLegalMemtileConnection(WireBundle srcBundle,
-                                               uint32_t srcChan,
+                                               int srcChan,
                                                WireBundle dstBundle,
-                                               uint32_t dstChan) const {
+                                               int dstChan) const {
   // Memtile north-south stream switch constraint
   // Memtile stream interconnect master South and slave North must have equal
   // channel indices
@@ -535,57 +533,57 @@ bool AIE2TargetModel::isLegalMemtileConnection(WireBundle srcBundle,
 void AIETargetModel::validate() const {
   // Every tile in a shimtile row must be a shimtile, and can only be one type
   // of shim tile.
-  for (uint32_t j = 0; j < columns(); j++) {
+  for (int j = 0; j < columns(); j++) {
     assert(!isMemTile(j, 0) && (isShimPLTile(j, 0) || isShimNOCTile(j, 0)) &&
            !isCoreTile(j, 0));
     assert(isShimPLTile(j, 0) ^ isShimNOCTile(j, 0));
   }
 
   // Every tile in a memtile row must be a memtile.
-  for (uint32_t i = 1; i < 1 + getNumMemTileRows(); i++)
-    for (uint32_t j = 0; j < columns(); j++)
+  for (int i = 1; i < 1 + (int)getNumMemTileRows(); i++)
+    for (int j = 0; j < columns(); j++)
       assert(isMemTile(j, i) && !isShimPLTile(j, i) && !isShimNOCTile(j, i) &&
              !isCoreTile(j, i));
 
   // Every other tile is a coretile.
-  for (uint32_t i = 1 + getNumMemTileRows(); i < rows(); i++)
-    for (uint32_t j = 0; j < columns(); j++)
+  for (int i = 1 + (int)getNumMemTileRows(); i < rows(); i++)
+    for (int j = 0; j < columns(); j++)
       assert(!isMemTile(j, i) && !isShimPLTile(j, i) && !isShimNOCTile(j, i) &&
              isCoreTile(j, i));
 
   // Looking North, buses must match
-  for (uint32_t i = 0; i < rows() - 1; i++)
-    for (uint32_t j = 0; j < columns(); j++)
+  for (int i = 0; i < rows() - 1; i++)
+    for (int j = 0; j < columns(); j++)
       assert(getNumSourceSwitchboxConnections(j, i, WireBundle::North) ==
              getNumDestSwitchboxConnections(j, i + 1, WireBundle::South));
   // Looking South, buses must match
-  for (uint32_t i = 1; i < rows(); i++)
-    for (uint32_t j = 0; j < columns(); j++)
+  for (int i = 1; i < rows(); i++)
+    for (int j = 0; j < columns(); j++)
       assert(getNumSourceSwitchboxConnections(j, i, WireBundle::South) ==
              getNumDestSwitchboxConnections(j, i - 1, WireBundle::North));
   // Looking East, buses must match
-  for (uint32_t i = 0; i < rows(); i++)
-    for (uint32_t j = 0; j < columns() - 1; j++)
+  for (int i = 0; i < rows(); i++)
+    for (int j = 0; j < columns() - 1; j++)
       assert(getNumSourceSwitchboxConnections(j, i, WireBundle::East) ==
              getNumDestSwitchboxConnections(j + 1, i, WireBundle::West));
   // Looking West, buses must match
-  for (uint32_t i = 0; i < rows(); i++)
-    for (uint32_t j = 1; j < columns(); j++)
+  for (int i = 0; i < rows(); i++)
+    for (int j = 1; j < columns(); j++)
       assert(getNumSourceSwitchboxConnections(j, i, WireBundle::West) ==
              getNumDestSwitchboxConnections(j - 1, i, WireBundle::East));
   // Edges have no connections
-  for (uint32_t j = 0; j < columns(); j++)
+  for (int j = 0; j < columns(); j++)
     assert(getNumSourceSwitchboxConnections(j, rows() - 1, WireBundle::North) ==
            0);
-  for (uint32_t i = 0; i < rows(); i++)
+  for (int i = 0; i < rows(); i++)
     assert(getNumSourceSwitchboxConnections(columns() - 1, i,
                                             WireBundle::East) == 0);
-  for (uint32_t i = 0; i < rows(); i++)
+  for (int i = 0; i < rows(); i++)
     assert(getNumSourceSwitchboxConnections(0, i, WireBundle::West) == 0);
 
   // FIFOS are consistent
-  for (uint32_t i = 0; i < rows(); i++)
-    for (uint32_t j = 0; j < columns(); j++)
+  for (int i = 0; i < rows(); i++)
+    for (int j = 0; j < columns(); j++)
       assert(getNumSourceSwitchboxConnections(j, i, WireBundle::FIFO) ==
              getNumDestSwitchboxConnections(j, i, WireBundle::FIFO));
 }
